Introduce guest (default) and system users
Guest is the default non-logged-in user. Makes it easier to store
activities for unauthenticated users and avoids nil users in the code.

System user is a special-purpose user, intended for whenever there's
a need to create records automatically (to leave a trace that "system"
created them). There's no specific use for it in this starter app, but
it's a good thing to have, just in case.